Deloopings and Eilenberg-Mac Lane spaces (#1079)
In this pull request we add:

- A definition of what it means for a pointed type to be deloopable
- A definition of what it means for an H-space to be deloopable
- A definition of Eilenberg-Mac Lane spaces

There were some universe level questions that one can ask about the type
of deloopings. One can show that the type of deloopings of `X` in
universe `l2` is equivalent to the type of deloopings of `X` in the
universe level of `X`. Showing the forward implication required me to do
some work on smallness of pointed types and higher groups, which is
included in this pull request.

Furthermore, the file about morphisms of H-spaces has been revamped, for
better formalization of the preservation of the coherence law. This was
needed for the notion of equivalence of H-spaces, which was prerequisite
to the notion of delooping of H-spaces.