Shifts and unshifts of concepts around sequential colimits (#1070)
This PR formalizes shifts of sequential diagrams, morphisms of
sequential diagrams, cocones of sequential diagrams, homotopies of
cocones of sequential diagrams, and unshifts of the last two. This
machinery is used to show that dropping any finite number of vertices
from the beginning of a sequential diagram gives an equivalent
sequential colimit.