Postulate components of coherent two-sided inverses for function extensionality and univalence (#1119)
Changes the postulates for funext and univalence such that there is
judgmentally only one converse map to `htpy-eq` and `equiv-eq`. The main
benefit, however, is that now `eq-htpy` and `eq-equiv` will appear with
their own names in Agda interactive mode, rather than as `pr1 (pr1
(funext ... ...))` and `pr1 (pr1 (univalence ... ...))`.

I leave it for potential future work to prove
`is-retraction-eq-(equiv|htpy)'` and `coh-eq-(equiv|htpy)'` rather than
postulate them, **_if_** we want this. Note that we could get away with
even fewer postulates if we really wanted to (see
[`TypeTopology/UF.Lower-FunExt`](https://www.cs.bham.ac.uk/~mhe/TypeTopology/UF.Lower-FunExt.html)).