Pin macOS version to one with Intel runners (#1131)
The `macOS-latest` runners are now macOS 14, and GitHub only provides
arm64 machines for those. The setup-agda action doesn't have prebuilt
binaries of Agda for that, so I'm pinning the macos version to 13.

I also switched the website deployment workflow to run on Ubuntu instead
of macos, since the Linux runners are consistently faster for us.