Pullbacks of synthetic categories (#1183)
This PR contains the axioms for pullbacks of synthetic categories
together with the needed infrastructure, such as cospans and cone
diagrams.
It seemed impossible to do this with the existing axioms, so I
postulated an additional axiom which states that given natural
isomorphisms α β γ δ (with appropriate domains and codomains) together
with 3-isomorphisms α≅γ and β≅δ, there is an isomorphism of respective
horizontal composites: β * α≅δ * γ. I suppose this axiom is validatd by
Makkai's principle.
Note also that some lines exceed the maximum character limit. I wasn't
sure how to avoid this since I tried to avoid ad-hoc renamings and `let`
statement.