The Archimedean property of `ℕ` (#1256)
This is an instance where I'm not certain if this is the ideal type for
the result.

In the longer term, these PRs are building up to the Archimedean
property for the rationals en route to addition on the reals. The choice
is whether to do an increasingly long chain of embeddings of `ℕ` in the
integers, integer fractions, and rationals -- one of my drafts has
`rational-fraction-Z (in-fraction-Z (int-N n))`, or to use
[`multiple-Ab`](https://unimath.github.io/agda-unimath/group-theory.multiples-of-elements-abelian-groups.html),
which defines natural multiplication for all contexts, at which point we
might just define the Archimedean property in greater generality [on
linearly ordered
groups](https://en.wikipedia.org/wiki/Archimedean_property#Definition_for_linearly_ordered_groups).
That approach would also require lots of lemmas proving `multiple-Ab`
works equivalently to other kinds of multiplication, but it's plausible.
(There also feels like a bit of tension about linear ordering, because
the usual total orders are defined on normal inequality, but most
sources I see on the Archimedean property use strict inequality. I don't
feel like I have a strong sense of whether that makes a difference.)

My sense is that the right choice is to do the embeddings, and then
separately prove that they are equivalent to `multiple-Ab`, but I'm open
to pushback on that.