Arithmetically located Dedekind cuts (#1257)
We define what it means for a cut to be arithmetically located, and show
that a cut that is arithmetically located and closed under the ordering
of the rationals is located.

Reviewing the goal of addition on the reals, the strategy is:

* the Archimedean property of the rationals is necessary to show
Dedekind cuts are arithmetically located
* that allows us to approximate two reals `x` and `y` within `epsilon/2`
* the sums of the lower and upper bounds on `x` and `y` are within
`epsilon`
* the sum is arithmetically located
* the sum is located and therefore a valid Dedekind cut

This seems to take forever to compile -- `make pre-commit` took almost
five minutes despite only compiling this one file -- but I have
optimistically attempted marking more properties of the rationals
`abstract` in the hope that will help. Any other advice for improving
compilation times would be strongly appreciated.

I also keep failing to find a pre-proven identity that shows that `a +
(b - a)` is the identity in an Abelian group. There are lots of other
identities that are just slightly different, with a different ordering
or parenthesization, but it seems like that probably exists and I'm just
not seeing it?