Literature overview for Introduction to Homotopy Type Theory (#1333)
The plan for this PR is to contain as much of Part 1 of the book as
possible (that's chapters 1 to 8).

Typos/potential oversights found during transcription:
- Page 51, Exercise 5.8 (b), fourth equation: `y * succ y` should be `x
* succ y`
- Page 60, Exercise 6.1 (b), multiplication is denoted by juxtaposition
instead of `·`
- Page 83, Definition 8.2.1, "collatz" is spelled in lowercase in the
definition but capitalized in Remark 8.2.2
- Page 86, Corollary 8.2.5, upper bounds of families are only defined
later in Definition 8.3.1
- Page 96, Exercise 8.8, proving decidable equality in sigma types uses
Hedberg's theorem (Theorem 12.3.5) and characterization of
identifications of dependent pair types (Theorem 9.3.4). Is there
supposed to be an alternative lower-tech proof?
- Page 96, Exercise 8.9 (b), showing decidable equality of dependent
functions sounds like it would need function extensionality (Axiom
13.1.3)