Similarity reasoning for real numbers (#1338)
It's small, but I'm putting it in its own PR because I think it might be
controversial. In particular, I don't know if you feel that we need an
entire large similarity relation notion -- and I'm uncertain how to make
the similarity reasoning syntax work if we did, at least without having
to repeat the similarity relation's name on every line. (I have some
vague ideas, but nothing certain.)

I feel pretty strongly that there should be some concise similarity
reasoning syntax for real numbers; there are going to be a lot of
equations and not a lot of ability to use equality reasoning.