Inflattices (#1319)
This builds on #1300 (and includes it), but I wanted to put it out as a
draft just to help work out what isn't in `order-theory` because it just
got skipped, versus what was deliberately left out, and I don't know the
right terms -- if any exist? -- for the join variants of e.g.
[locales](https://unimath.github.io/agda-unimath/foundation.locale-of-propositions.html).
In particular, as I'm building structures around the ordering of lower
and upper Dedekind reals from #1314, I find myself running into the
absence of an inflattice for
[powersets](https://unimath.github.io/agda-unimath/foundation.powersets.html#the-powerset-suplattice)
where a suplattice exists, and so on...