Multiplication on closed intervals of rational numbers (#1351)
A step towards multiplication on the real numbers. Shares a great deal
of code with #1335, which defines minimum and maximum on the rationals
(and all decidable total orders), but bounding the products of rational
numbers will allow us to show that the product of real numbers is
arithmetically located and therefore located.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>