Lipschitz-continuous functions between metric spaces (#1417)
This introduces the concept of *Lipschitz* functions between metric
spaces and proves that left/right multiplication of rational numbers is
Lipschitz.

More precisely, `metric-spaces.lipschitz-functions-metric-spaces`
introduces the following concepts and results:
  - Lipschitz constant for a function between metric spaces;
- Lipschitz functions between metric spaces (functions with some
Lipschitz constant);
  - constant functions are Lipschitz for all `α : ℚ⁺`;
- short functions are Lipschitz functions with Lipschitz constant equal
to 1;
  - Lipschitz functions are uniformly continuous;
  - the composition of Lipschitz maps is Lipschitz;
  - being a Lipschitz map is homotopy invariant;
  - Lipschitz maps preserve elements at bounded distance.

The metric space of Lipschitz functions is introduced in
`metric-spaces.metric-space-of-lipschitz-functions-metric-spaces`.

We also introduce a new concept in the `metric-spaces` module:
- `metric-spaces.elements-at-bounded-distance-metric-spaces`: the
equivalence relation of being at bounded distance in a metric space
(being in _some_ neighborhood).