Axioms of countable and dependent choice (#1463)
Countable choice has come up in #1458 as a necessary ingredient of the
classical completion of a metric space. I have hopes we can avoid ACC,
but the only constructive approach I've seen is HoTT's higher
inductive-inductive types, which I'm not sure work in Cubical Agda, let
alone unimath. I haven't managed to put that together -- but it seemed a
reasonable time to at least define these axioms and their relationships.