Refactor metric spaces (#1450)
Refactoring the metric space module (#1421)

- remove the notion of `Premetric` / `Premetric-Space`:
  
as we discussed, the term "premetric" has too many different meanings to
be reliably used in this library. What was called a `Premetric` is now a
`Rational-Neighborhood-Relation` and we only focus on (pseudo)metric
spaces.

- all (pseudo)metric spaces are saturated:

including saturation in the definition of pseudometric space is coherent
with the definition of [Richman premetric
spaces](https://ncatlab.org/nlab/show/Richman+premetric+space) and makes
a lot of things _easier_ (e.g.: the limit map is short, the metric space
of Cauchy approximations in a complete metric space is complete, etc.).

- simplify hierarchy in concepts (define things for for metric spaces
first, instead of the splitting premetric>pseudometric>metric):

although a lot of concepts only depend on the underlying neighborhood
relation, they are principally useful for (pseudo)metric spaces. The
main concepts (isometries, Cauchy approximations and their limits) are
defined for pseudometric and metric spaces; other concepts (short maps,
continuity) are only defined for metric spaces but we can easily lower
them to pseudometric spaces if we ever need them.

---------

Co-authored-by: Louis Wasserman <wasserman.louis@gmail.com>