Add submodules of left modules over rings
This PR adds subsets of left modules over rings. If such a subset
contains the zero element/vector, is closed under addition and
multiplication by a scalar from the ring, we call this subset a
submodule. The PR contains a proof that the intersection of a family of
submodules forms a submodule and a proof that a submodule has the
structure of a module.

Signed-off-by: Šimon Brandner <simon.bra.ag@gmail.com>