Bounded distance decompositions of metric spaces (#1482)
Given a metric space `A`, this PR introduces the bounded distance
decomposition of a metric space `A`: the metric space
```agda
  bounded-distance-decompsition-Metric-Space :
    Metric-Space (l1 ⊔ l2) (l1 ⊔ l2)
  bounded-distance-decompsition-Metric-Space =
    indexed-sum-Metric-Space
      ( set-bounded-distance-component-Metric-Space A)
      ( subspace-bounded-distance-component-Metric-Space A)
```
where 
- `set-bounded-distance-component-Metric-Space A` is the quotient set of
`A` by the equivalence relation of being at bounded distance;
- `subspace-bounded-distance-component-Metric-Space A` is the subspace
of elements in a given quotient class.
In each subspace `subspace-bounded-distance-component-Metric-Space A X`,
all elements are at bounded distance.

Any metric space is isometrically equivalent to its bounded distance
decomposition. The underlying equivalence is provided by the following
new results for quotient sets and equivalence classes:

```agda
module _
  {l1 l2 : Level} {A : UU l1} (R : equivalence-relation l2 A)
  where

  equiv-total-equivalence-class :
    Σ (equivalence-class R) (type-subtype ∘ is-in-equivalence-class-Prop R) ≃ A
  equiv-total-equivalence-class = [...]

  equiv-total-set-quotient :
    Σ ( set-quotient R)
       ( type-subtype ∘ is-in-equivalence-class-set-quotient-Prop R) ≃
    ( A)
 equiv-total-set-quotient = [...]
```