Commutative semigroups (#1494)
I probably need to add sums over finite types, not just the standard
finite type, too.

I would love to get away without this, but soon we will need maxima and
minima over finite families, and max and min are not monoids in general
(and certainly not for the rationals and reals where we need them).