Metric spaces induced by metric functions (#1520)
I do find myself wondering if it'd be better to define the metric space
of real numbers in terms of the distance function, rather than in terms
of the neighborhood shenanigans, though they are of course equivalent.
(I haven't bothered proving that yet.)

This of course gives locatedness and bounded distance for free.

(I also haven't yet proven that the distance function available to all
located metric spaces with all elements at bounded distance is the same
as the original distance function.)