Category of algebras of a theory (#1483)
For a signature `S` and `S`-theory `T`, there is a large precategory of
`T`-algebras and `T`-algebra homomorphisms. Even better, this (large)
precategory is a (large) category.

To define this category, this PR also characterizes the identity types
of `T`-algebras, as both homomorphisms which are also equivalences and
as equivalences which preserve operations, making good use of various
identity-type technology in the `foundation` namespace.