Disable `typecheck`, `link-check`, and `pre-commit` workflows on master (#1552)
The previous change to the `typecheck` workflow #1493 contained a
mistake that made it skip on commits to the master branch, which was
unintended. However, I believe that we in fact do not need to run it on
master, since it is already tested in the pull request branch, and
besides, the benchmarking workflow also verifies that master type
checks. Hence, this PR hides the workflow from the workflow agenda on
master. It does so by creating a new ci file.

This PR also renames the benchmarking/profiling workflow to "Clean build
and profiling of library".

I've still added an inclusion rule so that if we in the future wish to
run it on commits to a certain branch we only need to add an "on push"
entry and not edit the if clause for `typecheck`.