Intervals and interval arithmetic (#1497)
In progress work on #1485. Still working on, in particular, that given
intervals [a,b] and [c,d], that Minkowski multiplication is surjective
onto [min(ac,ad,bc,bd), max(ac,ad,bc,bd)]; it's definitely true but
seems like it'll require endless casework again. That said, at least it
is plausibly WLOG-able, unlike the width bound we needed for real
multiplication as far as I can tell...

Not ready, and especially waiting for the deletion of the _old_ closed
interval file, but posted in response to
https://github.com/UniMath/agda-unimath/issues/1485#issuecomment-3218529244
.