Organize repository files (#1500)
Tidies and organizes the repository files better. Also includes some
fixes I made along the way.

- Move Mdbook theme from `theme` to `website/theme`.
- Move mdbook source directory from `docs` to `book/src`.
- place `CONTRIBUTORS.md`, `MAINTAINERS.md` and `SUMMARY.md` directly
into the mdbook source directory. This means that the `clean` entrypoint
in the Makefile now also deletes these.
- Move `.md` docs to new `docs` folder, including `tables` as a
subdirectory. (Does not include `README.md`, `CONTRIBUTING.md`, or
`LICENSE.md`)
- Move config files that are not required to be in the root directory to
a new folder `config`.

- Header in `CONTRIBUTORS.md` file is now top-level
- Disable git metadata content for `STATEMENT-OF-INCLUSION.md`, and
rename it to `STATEMENT-OF-INCLUSIVITY` in line with its name in
`SUMMARY.md`.
- Remove the "generate main index" pre-commit hook. It should never fail
on the user side, and the description was wrong.
- Correct formatting of a `contributors.py` warning message.
- Makefile `clean` entrypoint now also deletes profiling temp files and
the pycache, and the website files can be cleaned separately with `make
clean-website`.
- Add description and author metadata to mdbook.
- Editorconfig now configures Makefiles to have tab indentation.
- Certain references were not capitalized properly.
- A bunch of broken cases for the GitHub edit button

Resolves #1061

---

- `master`: 7 top-level folders, 34 top-level files
- `reorg`: 7 top-level folders, 16 top-level files