Large monoids (#1587)
It's not needed for the _internal_ laws, but the large similarity
relation and raising of universe levels seems necessary both to match
the models of the prototypical large structures (the real numbers, but I
suppose unions and intersections of subtypes are both monoids as well),
to have useful operations available like natural number powers, and to
get small monoids from large ones.

Related to #1334 , as that started to grapple with what large algebraic
structures would look like.