Raise minimal supported Python version to 3.10 (#1560)
We used to support Python 3.8 because of Ubuntu 20.04, which doesn't
have LTS anymore. The successor, Ubuntu 22.04, comes with Python 3.10.
By upgrading we get tooling improvements --- we can update
`pre-commit-hooks` and `autopep8`, the latter of which was disabled IIRC
because the version we could use was buggy --- and we can also use new
Python features, such as pattern matching and type aliases.

~~There's no time pressure, so I'm keeping it as a draft for now, since
one nested Python dependency has recently been broken for Python 3.10 in
nixpkgs, but I submitted a PR
(https://github.com/NixOS/nixpkgs/pull/447787) to fix that, and
temporarily replicated the fix here (with the disadvantage of it not
being prebuilt, hence waiting for the PR to be merged).~~
I got rid of the broken dependency by removing a dependency on
`requests`. We were only using it to perform a GET requests and looking
at its `Link` header, which can be done quite easily with Python's
built-in `urllib`.

Note that I updated pre-commit/prettier, but it's not maintained
anymore, because pre-commit apparently broke something, and IIUC there's
now no officially supported way of running prettier in pre-commit 🤷.

Fixes #939