Reject `{{#concept}}` calls with empty fields (#1620)
Fixes #1390. The preprocessor now correctly rejects calls with empty
`WDID=`, `WD=""`, `Disambiguation=""`, and `Agda=`. It doesn't reject
`WD=` and `Disambiguation=` though, because those aren't "empty fields",
since they aren't recognized as fields in the first place.

The next time I need to touch this preprocessor I'm rewriting the
parsing. When I first introduced it it was intentionally very loose in
what it accepts and no documentation was written, because it was
supposed to be used by an external project, and it wasn't clear what
metadata we should allow in there, so we allowed everything (for example
Disambiguation was at first just something Egbert would write and it
wouldn't be processed by anything).

By now the external project is shut down and we've converged on what
metadata we use, so we can write a proper parser.