Metric quotients of pseudometric spaces (#1622)
This PR introduces the notion of **metric quotient** of a pseudometric
spaces: the metric space whose objects are quotient classes under the
similarity relation in the pseudometric spaces and rational neighborhood
relations given by neighborhoods of class-elements.

This is a metric space isometric to the original pseudometric space. In
case of metric spaces, this isometry is an isometric equivalence. Any
short map (resp. isometry) from a pseudometric space to a metric space
factors as a short map (resp. isometry) through the metric quotient.

The module
`cauchy-approximations-metric-quotients-of-pseudometric-spaces`
introduces a few results regarding Cauchy approximations in pseudometric
spaces and Cauchy approximations in their metric quotients.

Co-authored-by: Louis Wasserman <wasserman.louis@gmail.com>