Every point in a proper closed interval in the real numbers is an accumulation point (#1679)
...and the definition of what an accumulation point _is_, and so on.

This is intended as a key step to #1615; let me sketch it out.

Let `f` be a function from `[a,b]` to the real numbers, where `a < b`.
The derivative of a function at a point is -- well, defined another way,
but we can show that given a sequence `y` approaching `x` but remaining
apart from it, the derivative at `x` is equal to the unique limit of (f
y - f x)/(y - x) as y approaches x.

Such a sequence exists if and only if x is an accumulation point of the
space where the function is defined. This PR proves every point in a
proper closed interval is an accumulation point, so the value of any
derivative of `f` at `x` is unique, so all derivatives are homotopic, so
derivatives are unique and being differentiable is a proposition.